z is the new j, yo!
`z` is a fish function testing new Jump functionality that I call base
dir mode. `z` in restricts the jumps only to the descendants of a git repo.
This is useful in a monorepo.

The base dir in Jump can be anything really, but setting it to a VCS
repo is useful.